+++ src/auth/consts.js
@@ -1,0 +1,5 @@
+module.exports.OICD_ISSUER_REL = 'http://openid.net/specs/connect/1.0/issuer'
+module.exports.USER_ROLES = {
+  ADMIN: 'admin',
+  USER: 'user'
+}