+++ views/ap/ap.js
@@ -7,6 +7,8 @@ import ServerDataContext from './ServerDataContext'
 import Thread from './Thread'
 import ObjectView from './ObjectView'
 import EmojiLink from '../components/EmojiLink'
+import Admin from '../admin/Admin'
+import { useCheckAdmin } from './utils/useCheckAdmin'
 
 const mountNode = document.getElementById('app')
 ReactDOM.render(<Root />, mountNode)
@@ -24,13 +26,19 @@ function Root () {
     ...window._serverData,
     isInIframe: inIframe()
   })
+  const isAdmin = useCheckAdmin(dataContext.token)
+
   const taskbarButtons = []
+  if (isAdmin) {
+    taskbarButtons.push(<EmojiLink key='admin' emoji='princess' href='/admin' title='Administrator Settings' />)
+  }
   if (dataContext.loggedInUser) {
     taskbarButtons.push(<EmojiLink key='logout' emoji='end' href='/auth/logout' title='Logout' />)
   } else {
     // login button
     taskbarButtons.push(<EmojiLink key='login' emoji='passport_control' href='/auth/login' title='Log in' />)
   }
+
   useEffect(() => {
     if (!dataContext.loggedInUser) {
       return
@@ -63,6 +71,7 @@ function Root () {
           <Profile path='/u/:actor/*' taskbarButtons={taskbarButtons} />
           <Thread path='/s/:activityId' taskbarButtons={taskbarButtons} />
           <ObjectView path='/o/:objectId' taskbarButtons={taskbarButtons} />
+          <Admin path='/admin' taskbarButtons={taskbarButtons} />
         </Router>
       </ServerDataContext.Provider>
     </IntlProvider>